+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 1.0.0a6 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Fixing subscription bug
 
 
 1.0.0a5 (2020-12-22)